requirements.txt
aniso8601==3.0.2
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
mysql-connector-python
flask-mysql
flask-sqlalchemy
requests